Requested Fixes
- Split original Painkiller_Patch.hpp into their respective files (CfgMagazines.hpp, CfgVehicles.hpp, CfgWeapons.hpp)

- Fixed image pathing per HEMTT

- Removed redundant line in ACE_Medical_Treatment_Actions.hpp

- Updated config.cpp to include the new broken out files
