+++ addons/miscMedical/CfgWeapons.hpp
@@ -1,0 +1,14 @@
+// License Attribution req: UI image authored by user grmarcstock via Vecteezy.com
+
+class CfgWeapons {
+    class ACE_ItemCore;
+    class ACE_painkillers_Item: ACE_ItemCore {};
+    class potato_pkblister: ACE_painkillers {
+        scope = 2;
+        author = "AChesheireCat";
+        displayName = "Painkiller Blister Pack";
+        descriptionShort = "For physical pain only";
+        descriptionUse = "A blister pack containing individualized painkiller tablets, regularly issued in IFAKs";
+        picture = QPATHTOEF(miscMedical,ui\pkblister_ca.paa);
+    };
+};
