+++ addons/miscMedical/config.cpp
@@ -15,4 +15,7 @@ class CfgPatches {
 #include "CfgEventHandlers.hpp"
 #include "Naloxone_Patch.hpp"
 #include "Painkiller_Patch.hpp"
+#include "CfgMagazines.hpp"
+#include "CfgVehicles.hpp"
+#include "CfgWeapons.hpp"
 #include "ACE_Medical_Treatment_Actions.hpp"
